🚚 api: move controller to src context
