+++ api/src/certification/results/application/certification-results-controller.js
@@ -48,6 +48,31 @@ const getSessionResultsByRecipientEmail = async function (
     .header('Content-Disposition', `attachment; filename=${csvResult.filename}`);
 };
 
-const certificationResultsController = { getCleaCertifiedCandidateDataCsv, getSessionResultsByRecipientEmail };
+const getSessionResultsToDownload = async function (
+  request,
+  h,
+  dependencies = { tokenService, getSessionCertificationResultsCsv },
+) {
+  const token = request.params.token;
+  const { sessionId } = dependencies.tokenService.extractCertificationResultsLink(token);
+  const { session, certificationResults } = await usecases.getSessionResults({ sessionId });
+
+  const csvResult = await dependencies.getSessionCertificationResultsCsv({
+    session,
+    certificationResults,
+    i18n: request.i18n,
+  });
+
+  return h
+    .response(csvResult.content)
+    .header('Content-Type', 'text/csv;charset=utf-8')
+    .header('Content-Disposition', `attachment; filename=${csvResult.filename}`);
+};
+
+const certificationResultsController = {
+  getCleaCertifiedCandidateDataCsv,
+  getSessionResultsByRecipientEmail,
+  getSessionResultsToDownload,
+};
 
 export { certificationResultsController };
