+++ api/tests/unit/application/session/session-controller_test.js
@@ -54,47 +54,6 @@ describe('Unit | Controller | sessionController', function () {
     });
   });
 
-  describe('#getSessionResultsToDownload ', function () {
-    it('should return results to download', async function () {
-      // given
-      const i18n = getI18n();
-      const session = { id: 1, date: '2020/01/01', time: '12:00' };
-      const sessionId = session.id;
-      const fileName = `20200101_1200_resultats_session_${sessionId}.csv`;
-      const certificationResults = [];
-      const token = Symbol('a beautiful token');
-      const request = {
-        i18n,
-        params: { id: sessionId, token },
-        auth: {
-          credentials: { userId },
-        },
-      };
-      const dependencies = {
-        getSessionCertificationResultsCsv: sinon.stub(),
-        tokenService: {
-          extractCertificationResultsLink: sinon.stub(),
-        },
-      };
-      dependencies.tokenService.extractCertificationResultsLink.withArgs(token).returns({ sessionId });
-      dependencies.getSessionCertificationResultsCsv
-        .withArgs({
-          session,
-          certificationResults,
-          i18n: request.i18n,
-        })
-        .returns({ content: 'csv-string', filename: fileName });
-      sinon.stub(usecases, 'getSessionResults').withArgs({ sessionId }).resolves({ session, certificationResults });
-
-      // when
-      const response = await sessionController.getSessionResultsToDownload(request, hFake, dependencies);
-
-      // then
-      expect(response.source).to.deep.equal('csv-string');
-      expect(response.headers['Content-Disposition']).to.equal(`attachment; filename=${fileName}`);
-    });
-  });
-
   describe('#createCandidateParticipation', function () {
     const sessionId = 1;
     const userId = 2;
